feat: add a minimal example for parsing and validating a TOML-based plugin configuration
parse a TOML-based plugin configuration to JSON, using pydantic to validate the configuration client-side and then generating a JSON-string that can be sent to the server, for example to render the plugin configuration on the web UI.

An example TOML file is provided (input_params.toml). An ipython notebook (parse_toml_to_model.ipynb) can be executed to test the parsing and validating functionality. Add and change values inside the TOML file to see what effect this has on the validation step.